+++ app/src/main/java/com/example/firstcomposeproject/MainActivity.kt
@@ -4,6 +4,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
+import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Box
@@ -13,11 +14,17 @@ import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardColors
+import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.RectangleShape
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.example.firstcomposeproject.ui.theme.InstagramProfileCard
@@ -42,31 +49,25 @@ fun backgroundColor(column: Int, row: Int): Color {
 
 @Preview
 @Composable
-fun TimesTable() {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
+fun CardTest() {
+    Card(
+        shape = RoundedCornerShape(
+            topStart = 4.dp,
+            topEnd = 4.dp,
+        ),
+        colors = CardDefaults.cardColors(
+            containerColor = Color.Green,
+            contentColor = Color.White
+        ),
+        border = BorderStroke(
+            1.dp,
+            Color.Black
+        )
     ) {
-        for (i in 1 until 10) {
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .weight(1f)
-            ) {
-                for (j in 1 until 10) {
-                    Box(
-                        modifier = Modifier
-                            .fillMaxHeight()
-                            .weight(1f)
-                            .background(color = backgroundColor(column = i, row = j))
-                            .border(width = 1.dp, color = Color.Black),
-                        contentAlignment = Alignment.Center
-                    ) {
-                        Text(text = "${i * j}", color = Color.Black)
-                    }
-                }
-            }
-        }
+        Text(
+            modifier = Modifier.padding(8.dp),
+            text = "Hello World!"
+        )
     }
 }
