+++ app/src/main/java/com/example/firstcomposeproject/ui/theme/InstagramProfileCard.kt
@@ -1,5 +1,6 @@
 package com.example.firstcomposeproject.ui.theme
 
+import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -10,6 +11,10 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardColors
+import androidx.compose.material3.CardDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -21,22 +26,37 @@ import androidx.compose.ui.unit.dp
 @Composable
 fun InstagramProfileCard() {
 
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(8.dp),
-        horizontalArrangement = Arrangement.SpaceEvenly,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Box(
-            modifier = Modifier
-                .size(50.dp)
-                .background(color = Color.Yellow)
+    Card(
+        shape = RoundedCornerShape(
+            topStart = 4.dp,
+            topEnd = 4.dp
+        ),
+        colors = CardDefaults.cardColors(
+            containerColor = Color.White
+        ),
+        border = BorderStroke(
+            width = 1.dp,
+            color = Color.Black
         )
-        TwoBoxes()
-        TwoBoxes()
-        TwoBoxes()
+    ) {
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(8.dp),
+            horizontalArrangement = Arrangement.SpaceEvenly,
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Box(
+                modifier = Modifier
+                    .size(50.dp)
+                    .background(color = Color.Yellow)
+            )
+            TwoBoxes()
+            TwoBoxes()
+            TwoBoxes()
+        }
     }
+
 }
 
 @Composable
